Moved functions and made mkdir a global function
